feature/#1425_Add_filter_to_disable_settings_flush_rewrite_rules_when_authors_page_is_enabled_disabled

+++ src/core/Plugin.php
@@ -600,8 +600,14 @@ class Plugin
         $supported_post_types = Utils::get_enabled_post_types();
         register_taxonomy($this->coauthor_taxonomy, $supported_post_types, $args);
 
-        if (delete_transient('ppma_flush_rewrite_rules')) {
-            flush_rewrite_rules(true);
+        if (apply_filters('publishpress_authors_flush_rewrite_rules', PUBLISHPRESS_AUTHORS_FLUSH_REWRITE_RULES)) {
+            /**
+             * Flush rewrite rules to enable permalink update
+             * for when author profile is enabled/disabled
+             */
+            if (delete_transient('ppma_flush_rewrite_rules')) {
+                flush_rewrite_rules(true);
+            }
         }
     }
